+++ src/pages/index/index.vue
@@ -86,6 +86,11 @@
               </dt>
             </div>
           </dl>
+          <pagination-component
+            v-if="paginationSettings && paginationSettings.total && !isLoading && incidents.length && isPaginationAvailable"
+            :pagination-settings="paginationSettings"
+            @selected-page="getIncidentsData()"
+          />
         </div>
       </div>
     </div>