+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Incident Center Release Notes
 
+* **feature:** Pagination is added to the homepage ([CE-1569](https://jira.rfcx.org/browse/CE-1569))
+
 ## 0.0.1 (2021-??-??)
 
 * **bug:** Response assets are loaded only when user opens them ([CE-1515](https://jira.rfcx.org/browse/CE-1515))